make all fee-related configs f64

+++ lib/l1_sender/src/config.rs
@@ -11,13 +11,13 @@ pub struct L1SenderConfig<Input> {
     pub operator_pk: SecretString,
 
     /// Max fee per gas we are willing to spend (in gwei).
-    pub max_fee_per_gas_gwei: u64,
+    pub max_fee_per_gas_gwei: f64,
 
     /// Max priority fee per gas we are willing to spend (in gwei).
     pub max_priority_fee_per_gas_gwei: f64,
 
     /// Max fee per blob gas we are willing to spend (in gwei).
-    pub max_fee_per_blob_gas_gwei: u64,
+    pub max_fee_per_blob_gas_gwei: f64,
 
     /// Max number of commands (to commit/prove/execute one batch) to be processed at a time.
     pub command_limit: usize,
@@ -34,7 +34,7 @@ pub struct L1SenderConfig<Input> {
 impl<T> L1SenderConfig<T> {
     /// Max fee per gas we are willing to spend (in wei).
     pub fn max_fee_per_gas(&self) -> u128 {
-        self.max_fee_per_gas_gwei as u128 * (GWEI_TO_WEI as u128)
+        (self.max_fee_per_gas_gwei * GWEI_TO_WEI as f64) as u128
     }
 
     /// Max priority fee per gas we are willing to spend (in wei).
@@ -44,6 +44,6 @@ impl<T> L1SenderConfig<T> {
 
     /// Max fee per blob gas we are willing to spend (in wei).
     pub fn max_fee_per_blob_gas(&self) -> u128 {
-        self.max_fee_per_blob_gas_gwei as u128 * (GWEI_TO_WEI as u128)
+        (self.max_fee_per_blob_gas_gwei * GWEI_TO_WEI as f64) as u128
     }
 }
